requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0